rename: 컴포넌트들 네이밍 수정
