+++ src/components/Team/Progress.tsx
