+++ src/pages/[id]/index.tsx
@@ -8,7 +8,7 @@ import Button, {
   TextColor,
   TextSize,
 } from '@/components/common/Button/Button';
-import DropDown from '@/components/common/dropdown';
+import DropDown from '@/components/common/Dropdown';
 import Input from '@/components/common/Input';
 import { Modal } from '@/components/modal';
 import TaskLists from '@/components/TaskList/TaskLists';
